requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
bcolz==1.2.1
beautifulsoup4==4.6.3
bleach==2.1.4
bokeh==0.13.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cliff==2.8.2
cligj==0.4.0
cloudpickle==0.5.5
cmd2==0.9.4
colorama==0.3.9
configparser==3.5.0
constantly==15.1.0
cryptography==2.3.1
cryptography-vectors==2.3.1
cssselect==1.0.3
cupy==6.0.0
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
dask==0.19.1
decorator==4.3.0
defusedxml==0.5.0
descartes==1.1.0
dill==0.2.8.2
distributed==1.23.1
entrypoints==0.2.3
fastrlock==0.4
feather-format==0.4.0
feedparser==5.2.1
Fiona==1.7.13
future==0.18.2
geopandas==0.4.0
graphviz==0.9
heapdict==1.0.0
html5lib==1.0.1
hyperlink==17.3.1
idna==2.7
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isoweek==1.3.3
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
kafka-python==2.0.2
kaggle==1.5.6
kaggle-cli==0.12.13
kiwisolver==1.0.1
locket==0.2.0
lxml==4.2.5
MarkupSafe==1.0
matplotlib==2.2.3
MechanicalSoup==0.8.0
mistune==0.8.3
mizani==0.4.6
mkl-fft==1.0.4
mkl-random==1.0.1
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.1
munch==2.3.2
murmurhash==0.28.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.6.0
numexpr==2.6.6
numpy==1.15.0
oauthlib==3.1.0
olefile==0.46
opencv-python==3.4.3.18
packaging==17.1
palettable==3.1.1
pandas==0.23.4
pandas-summary==0.0.5
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
path.py==11.0.1
patsy==0.5.0
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
plac==0.9.6
plotnine==0.4.0
pluggy==0.7.1
preshed==1.0.1
prettytable==0.7.2
progressbar2==3.34.3
prometheus-client==0.3.0
prompt-toolkit==1.0.15
psutil==5.4.7
ptyprocess==0.6.0
py==1.6.0
pyarrow==0.10.0
pyasn1==0.4.4
pyasn1-modules==0.2.1
pycparser==2.18
Pygments==2.2.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyperclip==1.6.4
pyproj==1.9.5.1
PySocks==1.6.8
pytest==3.8.0
python-dateutil==2.7.3
python-slugify==4.0.0
python-twitter==3.5
python-utils==2.3.0
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
regex==2017.11.9
requests==2.19.1
requests-oauthlib==1.3.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
six==1.11.0
sklearn-pandas==1.7.0
sortedcontainers==2.0.4
spacy==2.0.12
statsmodels==0.9.0
stevedore==1.29.0
tables==3.4.4
tblib==1.3.2
termcolor==1.1.0
terminado==0.8.1
testfixtures==6.3.0
testpath==0.3.1
text-unidecode==1.3
thinc==6.10.3
toolz==0.9.0
torch==0.3.1.post2
torchtext==0.2.3
torchvision==0.2.1
toree==0.4.0
tornado==4.5.3
tqdm==4.24.0
traitlets==4.3.2
tweepy==3.9.0
tweet-parser==1.13.2
Twisted==18.7.0
ujson==1.35
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.1
wrapt==1.10.11
xx-ent-wiki-sm==2.0.0
zict==0.1.3
zope.interface==4.5.0
